AI/VisionAgents/VisionAgentsPythonPlugins/TADA_TTS/pyproject.toml
[project]
name = "tada-tts"
version = "0.1.0"
description = "TADA TTS plugin for Vision Agents - Hume AI speech synthesis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "vision-agents[deepgram,gemini,getstream]>=0.4.3",
    "vision-agents-plugins-tada",
    "hume-tada>=0.1.6",
    "descript-audio-codec>=1.0.0",
    "python-dotenv>=1.0",
]

[tool.uv.sources]
vision-agents-plugins-tada = { path = "plugins/tada", editable = true }

[tool.uv]
override-dependencies = [
    "protobuf>=6.31.1",
    "transformers>=4.57.1,<5",
]
